requirements.txt
anthropic>=0.40.0
instagrapi>=2.3.0
yt-dlp>=2024.1.1
Pillow>=10.0.0
APScheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
